+++ src/app/components/content-provider-info.js
@@ -1,7 +1,7 @@
 /**
- * Popup with information about provider
+ * Content of popup with information about provider
  * 
- * @module components/provider-info-popup
+ * @module components/content-provider-info
  * @author Agnieszka Warchoł
  * @copyright (C) 2022 ACK CYFRONET AGH
  * @license This software is released under the MIT license cited in 'LICENSE.txt'.
@@ -10,10 +10,12 @@ import Component from '@ember/component';
 import I18n from 'onedata-gui-common/mixins/components/i18n';
 
 export default Component.extend(I18n, {
+  classNames: ['content-provider-info'],
+
   /**
    * @override
    */
-  i18nPrefix: 'components.providerInfoPopup',
+  i18nPrefix: 'components.contentProviderInfo',
 
   /**
    * @virtual