+++ CHANGELOG.md
@@ -6,6 +6,7 @@ CHANGELOG
 
 ### Latest changes
 
+* VFS-6803 Added Oneprovider info popup
 * VFS-6397 Removed redundant bower dependencies
 * VFS-8617 Removed usages of ember-invoke-action
 * VFS-8574 Updated backend errors translations